requirements.txt
oracledb
pika
ccard
python-dotenv